requirements.txt
valohai-utils>=0.1.12,<0.2
numpy>=1.22.1,<1.23
